It was impossible to determine if keyboard interactive authentication was interactively cancelled by the user when using the blocking authenticate() method. Due to the nature of KBI, the same solution as used for PasswordAuthentication (returning null, which throws an exception) will not work here, so I implemented this a different.
ClientAuthenticator now has an isCancelled() and cancel() methods. Both PasswordAuthenticator and KeyboardInteractiveAuthenticator will now call this when either password is null, or KeyboardInteractiveCompletor's cancel() method is called. Then, isCancelled() is tested for in the blocking authenticate() method. If it is true, an exception is thrown.

Non block could of course also use isCancelled(), and this method matches the pattern used for isDone(), is Failure() etc.

@ludup please review.