Do not recreate the executor on shutdown
If the ssh server is stopped or closed twice, the shutdown can end up
in an infinite loop that overloads the heap. This is due to it
recreating the executor service and adding a new shutdown hook
after the first stop or close set the executor to null. This
commit uses the executor static field value directly in shutdown
instead of getExecutorService to avoid the infinitely recreated
shutdown hooks.