`SftpFile` no longer will hold any copy of `SftpFileAttributes`, it will instead always retrieve from the network  on calling `SftpFile.getAttributes()` (which it *might* have done anyway depending on usage). Setting should now preferably done via the `SftpFile.setAttributes()` object too, which delegates to `SftpChannel`.
I have gone through all of our usages, and checked this won't harm performance, and it appears not. In all places I could discover, it would not make any difference as the stat call was always being done before constructing the `SftpFile` object. In fact it will help in many cases too, as attributes will only be retrieved when they are needed.

I would like to try to work towards the following set-in-stone architecture ..

 * `SftpChannel` - private API
 * `SftpClient` - lowest level public API
   * `SftpHandle` - Publicly immutable (has a closed state though)
   * `SftpFile` - Entirely immutable (it just about is now bar the unused "properties" it carries)
   * `SftpFileAttributes` - Mutable carrier of attributes that may be retrieved from stored