Introduces `SftpFileAttributesBuilder` and a redesigned but currently mostly API compatible `SftpFileAttributes`. It has been prepared to be made immutable at version 3.2.0, when the deprecated methods and constructors will be removed.
`SftpFile` continues to hold a copy of SftpFileAttributes, which may be refreshed or set. The recommended method of doing so will be on the SftpFile object itself. SftpFile itself will be mostly immutable. It will represent a unique path which will never change, and the immutable attributes that may change.

This commit also introduces an alternative method of creating links and symbolic links and deprecates the current way. This was done because I found that it is impossible to create relative links (totally valid thing to do).