Guard against a handle that has no `SftpFile` associated with it, and just use `<fileless-handle>` as the filename. Being as this is only for logging purposes, that will suffice for now.
Use of `getBestHandle()` is to be discouraged for this reason, the intention was always to remove it at 3.2.0.

All the SFTP NIO tests are passing again except the 3 remote-to-remote tests. I am not sure if its all these recent SFTP changes that broke this those, so that will need to be checked first